Various little changes (see desc.)
- Black formatting
- Move things into a `abcd` folder
- Add setup.py
- Convert command-line arguments to Dataclasses using simple-parsing
