+++ .gitignore
@@ -8,3 +8,4 @@ results/
 __pycache__/
 *.py[cod]
 *$py.class
+*.egg-info
